libct/cgroups: move RemovePath from fs2
This is to be used by RemovePaths.

(cherry picked from commit 3f14242e0a3b01a18621aa5efd91f87d915bd947)
